Added function to compute Gram matrix as a style representation

+++ src/methods.py
@@ -103,11 +103,11 @@ def cv2_image_to_tensor(img):
     Returns
     - tensor: 3-color RGB tensor.
     """
-
+    
     # Convert the BGR image to RGB
     rgb_img = cv2.cvtColor(img, cv2.COLOR_BGR2RGB)
     # Convert the RGB OpenCV image to a tensor.
-    tensor = tf.convert_to_tensor(rgb_img)
+    tensor = tf.convert_to_tensor(rgb_img, dtype='int32')
     return tensor
